Request and store payout when activating lease
As the first step of implementing the payout (i.e. Royalty)
functionality, we store the payout info given by the NFT contract when
transfering the NFT to our contract. This info will be later used to
split the rent payment to all relavant parties when the NFT got claimed
back.

It also replaced the nft_transfer callback with a private contract
function called `activate_lease`. This is a better way to handle
multiple async operation (i.e. promises) in a transcational way.