Make payment according to the payout info. (#12)
Added a `resolve_claim_back` private function to handle the steps after
transfering the NFT back to the owner: 1. make the payment to all
relevant parties, 2. clean up the lease record.

Also make the `internal_transfer_near` return a promise. It can be
useful in the future we want to make sure such transfer indeed goes
through with no issue.