[NR67][NR42]Add index to support get_borrower & add lease state check to get_borrower (#16)
### Context
Previously, get_borrower iterates the whole lease map to find the
matching borrower. It is not gas efficient
The method is also not checking lease status, where only active lease
should be counted

### Solution & Code change
- A new index _lease_id_by_contract_addr_and_token_id_ is added, which
can then use lease_map to get the matching borrower. This design is
easier to be extended in the future when other lease info needs to be
queried given (contract_addr, token_id).
- Lease state check added in the method when matching lease is found
- Updated the method name to be more descriptive
- Adjusted code to maintain the new index at insertion and removal
- Added unit tests