Use U128 for the price field in LeaseCondition.
Passing u128 to the frontend causes the floating point number
precision problem. Using U128 will fix it.

In general, U128 is a more friendly type for input and output. The only
drawback is that it take a little more storage space. We can optimise it
in the future if needed.